Put every method in sub file
